Create NeoStrip.ino
A way to use both the on board neoPixels and a neo strip with Circuit Playground - we searched and tried a lot of code and finally came on to this code. This has been tested and it works! Using Adafruit_CPlay_NeoPixel documentation is still a little hard to find about this.